requirements/frozen.txt
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.1
asttokens==3.0.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
CacheControl==0.14.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.13.4
cryptography==46.0.5
daphne==4.2.1
decorator==5.2.1
distlib==0.4.0
dj-database-url==2.3.0
Django==5.2.12
django-cors-headers==4.9.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
factory_boy==3.3.3
Faker==24.14.1
filelock==3.25.2
firebase-admin==6.9.0
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.25.0
google-cloud-storage==3.9.0
google-cloud-translate==3.24.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==22.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
identify==2.6.17
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
ipython==8.38.0
jedi==0.19.2
kombu==5.6.2
matplotlib-inline==0.2.1
msgpack==1.1.2
nodeenv==1.10.0
packaging @ file:///home/task_176104874243446/conda-bld/packaging_1761049080023/work
parso==0.8.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.9.4
pluggy==1.6.0
pre-commit==3.8.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psycopg==3.3.3
psycopg-binary==3.3.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
pytest==8.4.2
pytest-cov==5.0.0
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-discovery==1.1.3
PyYAML==6.0.3
redis==5.3.1
requests==2.32.5
resend==2.23.0
ruff==0.15.6
sentry-sdk==2.54.0
service-identity==24.2.0
setuptools==80.10.2
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
virtualenv==21.2.0
wcwidth==0.6.0
wheel==0.46.3
zope.interface==8.2
